Rename folders, fix "undefined" libraryName, fix Webpack merge strategy
