+++ ui.apps/src/main/.babelrc
@@ -1,3 +1,3 @@
 {
-    "extends": "./webpack/.babelrc"
+    "extends": "./webpack.core/.babelrc"
 }
