+++ ui.apps/src/main/webpack.core/.babelrc
