+++ ui.apps/src/main/webpack.core/internals/jest.config.js
@@ -1,6 +1,6 @@
 const path = require('path');
 const merge = require('merge')
-const CONFIG = require('./../../config');
+const CONFIG = require('./../../webpack.project');
 
 const JEST_DEFAULT = {
   "roots": [
