+++ ui.apps/src/main/webpack.core/package.json
@@ -35,7 +35,8 @@
     "postcss-loader": "^2.0.6",
     "sass-loader": "^6.0.6",
     "stylelint": "^7.13.0",
-    "webpack": "^3.6.0"
+    "webpack": "^3.6.0",
+    "webpack-merge": "^4.1.1"
   },
   "dependencies": {
     "babel-polyfill": "^6.23.0"
